Title: 실수 입력받아 계산 2, Time: 147ms, Memory: 24MB, Status: Passed - CodeTree

+++ 240312/실수 입력받아 계산 2/calculation-by-inputting-a-real-number-2.py
@@ -1,4 +1,4 @@
-a = int(input())
+a = float(input())
 b = a + 1.5
 
 print(f"{b:.2f}")
